interactive_RAG/requirements.txt
docling>=2.5.0
faiss-cpu>=1.8.0
PyMuPDF>=1.24.0
streamlit>=1.35.0
streamlit-image-coordinates>=0.4.0
openai>=1.30.0,<2.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
Pillow>=10.0.0
# RAG notebook dependencies
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-community>=0.4.0
jupyter>=1.0.0
ipykernel>=6.0.0
